requirements.txt
# MOAT Orchestration Backend Requirements

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Async
aiofiles>=23.0.0
httpx>=0.25.0

# Data Processing
numpy>=1.24.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.0  # Direct PostgreSQL connection (bypasses PostgREST cache)

# Logging
structlog>=23.0.0

# Testing dependencies moved to requirements-dev.txt
# Install dev deps with: pip install -r requirements-dev.txt

